% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapiverse-package.R
\docType{package}
\name{hapiverse-package}
\alias{hapiverse}
\alias{hapiverse-package}
\title{hapiverse: Easily Install and Load the 'hAPIverse'}
\description{
The 'hapiverse' is a collection of packages that work
    together to provide programmatic access to the European Bioinformatics
    Institute (EMBL-EBI) REST API servers. Currently, this meta-package
    is designed to make it easy to install and load the gwasrapidd and
    quincunx packages <https://maialab.org/hapiverse>.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte.scientist@gmail.com} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item Center for Health Technology and Services Research (CINTESIS), University of Algarve [copyright holder, funder]
}

}
\keyword{internal}
