% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{hapiverse_conflicts}
\alias{hapiverse_conflicts}
\title{Conflicts between the hapiverse and other packages}
\usage{
hapiverse_conflicts()
}
\description{
Conflicts are namespace collisions (i.e. functions that have the same name,
but are provided by different packages). This function lists all the
conflicts between packages in the hapiverse and other packages that you have
loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
hapiverse_conflicts()
}
