% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hapiverse_dependencies}
\alias{hapiverse_dependencies}
\title{List all package dependencies in the hapiverse}
\usage{
hapiverse_dependencies(include_self = TRUE)
}
\arguments{
\item{include_self}{Include hapiverse in the list?}
}
\description{
This function lists all direct package dependencies of hapiverse.
}
\examples{
hapiverse_dependencies()
}
