% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hapiverse_packages}
\alias{hapiverse_packages}
\title{List all packages in the hapiverse}
\usage{
hapiverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include hapiverse in the list?}
}
\description{
List all packages in the hapiverse
}
\examples{
hapiverse_packages()
}
