% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{package_version}
\alias{package_version}
\title{Pretty package version string}
\usage{
package_version(x)
}
\arguments{
\item{x}{The name of an installed package, a character string. NB: the
function is not vectorised on this argument.}
}
\value{
A stylised character string of the package version.
}
\description{
This function takes the name of a package, finds its version, and generates a
string where the fourth and subsequent fields are coloured in red (uses the
crayon package).
}
\keyword{internal}
