% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{hapiverse_packages}
\alias{hapiverse_packages}
\title{hapiverse packages}
\usage{
hapiverse_packages()
}
\value{
A character vector of the names of packages included in the hapiverse.
}
\description{
This function returns the packages that are part of the hapiverse. Contrast
with \code{\link[hapiverse]{hapiverse_dependencies}}, which returns all
package dependencies.
}
